pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deep-analysis-agent"
version = "0.4.0.dev0"
description = "Deep Analysis agent — Windows tray client for MTGO match analytics"
license = { text = "MIT" }
requires-python = ">=3.12"
readme = "README.md"

dependencies = [
    "pystray>=0.19",
    "watchdog>=4.0",
    "httpx>=0.27",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "structlog>=24.2",
    "Pillow>=10.0",
    "pywin32>=306; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
deep-analysis-agent = "deep_analysis_agent.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["pystray", "pystray.*", "win32event", "win32api", "winerror", "win32con"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
